feat!: Remove Valkey and change cache destination to local files
While Valkey seems like it would be a better solution, it has no
streaming API. This meant the entire response had to be buffered,
resulting in ~60 MB of memory used per request.

I may add Valkey back in the future and chunk the response to get the
best of both worlds.